# Conflicts:
#	frontend/src/user/pages/ProjectDashboard.jsx
#	frontend/src/user/pages/UserDashboard.jsx

+++ backend/src/main/java/com/codecool/tasx/controller/dto/reward/RewardCreateRequestDto.java
@@ -3,5 +3,5 @@ package com.codecool.tasx.controller.dto.reward;
 import java.util.List;
 
 public record RewardCreateRequestDto(Long id, Long userId, Long companyId, List<Long> taskIds, String name,
-                                     String description, int pointCost) {
+                                     String description, long pointCost) {
 }
